test: date & dateHelpers 100%
